RDKEMW-6566: Change display-offset param to 64-bit
The change is necessary because a 32-bit offset is not always enough.
The underlying libraries can easily handle 64-bit, but unfortunately
this interface was created with 32-bit.
We have control of both clients that use this function, and the change
is desired by one and unused by the other, so we believe it's safe.